Fixing minor bugs

+++ src/main/java/org/openjfx/model/Weather.java
@@ -1,7 +1,6 @@
 package org.openjfx.model;
 
 import javafx.beans.property.*;
-import javafx.beans.value.ObservableValue;
 import javafx.scene.image.Image;
 
 import static java.lang.Math.round;
@@ -10,12 +9,12 @@ public class Weather {
 
     private final SimpleStringProperty cityName;
     private final SimpleStringProperty conditions;
-    private final SimpleObjectProperty<Double> tempInCelsius;
+    private final SimpleObjectProperty<Integer> tempInCelsius;
     private final SimpleStringProperty dateTime;
     private final SimpleObjectProperty<Image> currentConditionsImage;
     private final SimpleStringProperty dayOfTheWeek;
 
-    public Weather(String cityName, String conditions, double tempInCelsius, String dateTime, Image currentConditionsImage, String dayOfTheWeek) {
+    public Weather(String cityName, String conditions, int tempInCelsius, String dateTime, Image currentConditionsImage, String dayOfTheWeek) {
         this.cityName = new SimpleStringProperty(cityName);
         this.conditions = new SimpleStringProperty(conditions);
         this.tempInCelsius = new SimpleObjectProperty<>(tempInCelsius);
@@ -28,7 +27,7 @@ public class Weather {
         return this.cityName.get();
     }
 
-    public double getTempInCelsius() {
+    public int getTempInCelsius() {
         return round(this.tempInCelsius.get());
     }
 
@@ -44,7 +43,7 @@ public class Weather {
         return this.currentConditionsImage.get();
     }
 
-    public SimpleObjectProperty<Double> getTempInCelsiusProperty(){
+    public SimpleObjectProperty<Integer> getTempInCelsiusProperty(){
         return this.tempInCelsius;
     }
